Refactor: Button.component 컨벤션에 맞춘다
