+++ src/components/common/index.ts
@@ -1,2 +1,2 @@
 export { default as Input } from './Input/Input';
-export { default as Button } from './Button/Button';
+export { default as Button } from './Button/Button.component';
